Changes to vchan blocking semantics
The previous semantics transmitted the "state" of a given vchan buffer up to the linux vm
e.g. Sending the code VCHAN_BUF_FULL

This results in overwrites when the buffer is not full but close to being full, and is generally a clunky situation.

The semantics are changed such that the linux vm vchan driver always knows how much data is ready to be read and how much buffer space is left, to catch such situations.