Allow the VMM to map frames belonging to other components into the VM's address space.
This is a hack for a demo showing how security can be broken if the VM is not properly isolated from other components.
This is a bit fragile in that it relies on attributes (num_extra_frame_caps and extra_frame_map_address) to be defined in the .camkes.
In order to use this feature, set the attributes in vm.camkes (number of frames, and where to map it in the VM's physical address apce (IPA)).
Then build.
Then edit the vm.cdl file to add desired frames in the vm's cnode (skip one slot after the last used slot).
Then build again and run.
In Linux the frames should be available at the chosen address accessible through /dev/mem.