Use custom IO mapper in the VM
Previously the VM used the libsel4platsupport IO mapper to map in
device pages. The libsel4platsupport mapper now makes assumptions
on devices being frames that can be allocated from an allocator.
This does not match with the VM, which has its resources as frames
in the simple interface, so we now create our own IO mapper to
handle this.