VM: Alloc frame for guest image prior to copy
Added the function 'maybe_create_frame_at'. This will attempt to
allocate a frame at a given address if it doesn't exist in the
guest vm's vspace prior. This is done before calling
'vm_guest_vspace_touch' when installing a vm image since the
function assumes the frame for the image has already been
allocated and mapped.