templates: Added new template seL4VMDTBPassthrough
Added a new template/interface that can be used in a camkes
arm vm app. seL4VMDTBPassthrough allows users to define a series
of dtb nodes that will in turn be converted into a series of
untyped memory and irq objects. This can then be used for passing
the devices through to the guest vm. This is an alternative
interface that can be used as opposed to specifying two
independent lists of untypes and irqs.