Update ping_client to use the new virtqueue lib
Introduces the necessary changes for the ping client to use the new
libvirtqueue with multi-entry rings and scatter buffer logic.

The camkes wrapper doesn't allocate the memory for the virtqueue objects
anymore so declare them as static data instead of pointers.

The camkes wrapper offers iterator functions to scatter/gather the
buffers, so the ping client now makes use of those.

The polling functions are different/simpler and cannot fail so
refactored some code there as well.

And finally the shared memory has been extended to hold the rings and
the descriptor table.