VM: Process unhandled memory faults
Processes unhandled memory faults from libsel4vm through the
callback interfaces and handles the exit if it can.
In particular the map on demand policy that originally
existed in libsel4vm has been ported to the vmm. Currently the VMM
will attempt to perform a map on demand for every mmio exit the
library forwards to it.