vm_cross_connector: Add custom crossvm device tree
Added a custom exynos5422 device tree with a modified pci device
node. As there a multiple crossvm devices on the PCI host sharing
irq '92'/INTA in this application, we create irq maps for pci
device ids 1-5. This is currently a static solution that should be
later replaced with more dynamic generation methods.