libvirtio: Dynamically allocate ioport ranges
We remove the static definitions of virtio ioport addresses in
favour of using new functionality introduced in
libsel4vmmplatsupport. The revisions in libsel4vmmplatsupport allows
for more dynamic allocation of ioport addresses. The advantage in
using this is that it allows us to program multiple virtio pci
devices without worrying about overlap.