VM: Refactored virtqueue net module to use vswitch
Refactored the virtio net virtqueue driver to use the vswitch
library. This extends the driver to be able to manage multiple
connections to different components and VM's. This now expects
each connection to be associated with a MAC address in order
to appropriately manage the different connections a VM can be in
LAN with.