virtio_con: use seL4_DebugPutChar if Domains
If the domain scheduler is being used, use seL4_DebugPutChar instead of
RPC calls to the serial server. This is because RPC calls to the serial
server will get blocked until a domain switch occurs and hang the VM
while this occurs. DebugPutChar should likely be replaced with a serial
server connection that doesn't use an RPC call for each character.